fix(rebase): resolve cli.rs refactor conflicts from main's cli/run.rs extraction
Main extracted CliMode/GlobalFlags/run() to cli/run.rs after this branch diverged.
Re-apply RAG v1 additions to the correct split files:
- cli/args.rs: add SimilarIncidents/AskHistory/IncidentContext variants + 3 new arg structs
- cli/run.rs: add dispatch arms for the 3 new AiCommand variants
- cli.rs: add parse_ai_similar/ask_history/incident_context fns + print fns + response imports
