+++ src/cli/args.rs
@@ -38,6 +38,9 @@ pub(crate) enum AiCommand {
     Doctor(AiDoctorArgs),
     WatchStatus(OutputArgs),
     SmokeWatch(OutputArgs),
+    SimilarIncidents(AiSimilarArgs),
+    AskHistory(AiAskHistoryArgs),
+    IncidentContext(AiIncidentContextArgs),
 }
 
 #[derive(Debug, Clone, PartialEq, Eq)]
@@ -386,6 +389,42 @@ pub(crate) struct AiErrorsArgs {
 pub(crate) struct AiPruneCheckpointsArgs {
     pub missing_only: bool,
     pub dry_run: bool,
+    pub limit: Option<u32>,
+    pub json: bool,
+}
+
+#[derive(Debug, Clone, Default, PartialEq, Eq)]
+pub(crate) struct AiSimilarArgs {
+    pub query: String,
+    pub hostname: Option<String>,
+    pub app_name: Option<String>,
+    pub severity_min: Option<String>,
+    pub from: Option<String>,
+    pub to: Option<String>,
+    pub window_minutes: Option<u32>,
+    pub limit: Option<u32>,
+    pub json: bool,
+}
+
+#[derive(Debug, Clone, Default, PartialEq, Eq)]
+pub(crate) struct AiAskHistoryArgs {
+    pub query: String,
+    pub hostname: Option<String>,
+    pub app_name: Option<String>,
+    pub from: Option<String>,
+    pub to: Option<String>,
+    pub limit: Option<u32>,
+    pub json: bool,
+}
+
+#[derive(Debug, Clone, Default, PartialEq, Eq)]
+pub(crate) struct AiIncidentContextArgs {
+    pub from: String,
+    pub to: String,
+    pub hostname: Option<String>,
+    pub app_name: Option<String>,
+    pub query: Option<String>,
+    pub severity_min: Option<String>,
     pub limit: Option<u32>,
     pub json: bool,
 }
