+++ src/compose.rs
@@ -13,7 +13,7 @@ pub use format::{ensure_doctor_ready, mcp_projection, redact_sensitive};
 pub use mutation::ComposeService;
 pub use runner::ProcessRunner;
 pub use types::{
-    CommandOutput, CommandRunner, ComposeDiagnostic, ComposeCommandResult, ComposeDefaults,
+    CommandOutput, CommandRunner, ComposeCommandResult, ComposeDefaults, ComposeDiagnostic,
     ComposeDryRun, ComposeInvocation, ComposeMcpDiagnostic, ComposeMcpStatus, ComposeMutation,
     ComposeOwnershipState, ComposeRuntimeState, ComposeStatus, ComposeTarget, ComposeTargetSummary,
     ContainerInfo, DiagnosticSeverity, DockerInspect, ListenerInfo, MountInfo, MutationOptions,
@@ -23,13 +23,16 @@ pub use types::{
 
 // Test-only re-exports of private items accessed via `use super::*` in compose_tests.rs.
 #[cfg(test)]
-pub(crate) use docker::{container_info_from_inspect, DockerUnavailableError, ss_output_has_listener, systemd_status_from_output};
+pub(crate) use docker::{
+    container_info_from_inspect, ss_output_has_listener, systemd_status_from_output,
+    DockerUnavailableError,
+};
 #[cfg(test)]
 pub(crate) use format::{status_from_target, unresolved_status};
 #[cfg(test)]
 pub(crate) use mutation::{
-    DIAG_DOCKER_UNAVAILABLE, DIAG_SYSTEMD_CHECK_FAILED, DIAG_TARGET_UNRESOLVED,
-    target_from_container, unresolved_code,
+    target_from_container, unresolved_code, DIAG_DOCKER_UNAVAILABLE, DIAG_SYSTEMD_CHECK_FAILED,
+    DIAG_TARGET_UNRESOLVED,
 };
 
 #[cfg(test)]
